feat: break streaks, rewards and activities into components
