+++ src/Redux/Store/store.ts
@@ -4,7 +4,7 @@ import { AccountSlice } from "../../Redux/Slices/accountSlice";
 import { PaymentSlice } from "../../Redux/Slices/paymentSlice";
 import { RecentActivitiesSlice } from "../../Redux/Slices/recentActivitiesSlice";
 import { RewardSlice } from "../../Redux/Slices/rewardSlice";
-import { StrealDetailSlice } from "../../Redux/Slices/streakDetailSlices";
+import { StreakDetailSlice } from "../../Redux/Slices/streakDetailSlices";
 import { StreakSlice } from "../../Redux/Slices/streakSlice";
 
 import storage from "redux-persist/lib/storage";
@@ -29,7 +29,7 @@ const store = configureStore({
       .concat(PaymentSlice.middleware)
       .concat(RecentActivitiesSlice.middleware)
       .concat(RewardSlice.middleware)
-      .concat(StrealDetailSlice.middleware)
+      .concat(StreakDetailSlice.middleware)
       .concat(StreakSlice.middleware),
 });
