+++ src/index.scss
@@ -290,6 +290,10 @@ a:active {
   align-items: baseline;
 }
 
+.ai-start {
+  align-items: flex-start;
+}
+
 .roboto-medium {
   font-family: RobotoMedium;
 }
